fix(shaders): resolve GLSL ES 300 strict-compile errors (13 DarkWindow + 1 wallpaper) (#26)
* fix(shaders): resolve GLSL ES 300 strict-compile errors across DarkWindow + wallpaper catalog

Drove validate-darkwindow-shaders.sh from 126/139 passing → 139/139, and
fixed the one wallpaper shader that failed raw glslangValidator.

Categories of fix (all real, not stylistic):

- Type strictness (int literal in a float-only overload context — ES 300
  does not auto-promote int→float like desktop GL): picom-vignette,
  cursor_blaze, in-game-crt (3 sites), in-game-crt-cursor (3 sites),
  inside-the-matrix, crt-shader-with-chromatic-aberration-glow-scanlines-
  dot-matrix
- Packed-bayer storage: dither.glsl stored 16-bit hex literals in a
  float[] (invalid since the constants don't convert cleanly to float);
  rewired to int[] and dropped the cast used on readback.
- Sampler-vs-function confusion: glitchy + computer-glitchy-4 passed
  x_Texture (a wrapper function) to textureLod/texelFetch, which expect
  a sampler2D; switched to the underlying `tex` sampler.
- Reserved-word collision: picom-cube's `half` → `halfSize`;
  wallpaper-shaders/hex-matrix.frag's `active` → `activeCell`.
- Typo: manga_slash `norm()` → `normalize()`.
- Shadertoy port: matrix.glsl referenced iChannel1 + iChannelResolution,
  which Hypr-DarkWindow does not expose. Replaced with a procedural
  hash22() 2D pseudo-noise; also fixed direct mutation of the x_PixelPos
  varying (ES 300 forbids; uses a local copy).
- Entrypoint adapter: matrix_rain.glsl defined only the Shadertoy
  mainImage(); added a thin windowShader() wrapper that forwards
  x_PixelPos so the DarkWindow validator template and runtime can find
  the expected entrypoint.

Surface: 14 files total (13 DarkWindow + 1 wallpaper). No behavioral
changes — all edits preserve each shader's visual intent.


* fix(hyprland): bump touchpad sensitivity to 0.75

Per-device override for elan06fa:00-04f3:327e-touchpad.
Was -0.5899375 (~0.5x libinput default); 0.75 maps to
~2x default — roughly 4x the prior speed.


* chore(hyprdynamicmonitors): pin fallback to native eDP-1 layout

Set the fallback profile to drive eDP-1 at its native 2560x1600@240
with scale 1, and drop unknown monitors to preferred at scale 1
instead of 2.


* chore(audit): allowlist quickshell-menu-data (consumed via QS IPC)

quickshell-menu-data.sh is invoked by quickshell/modules/MenuOverlay.qml
through the Quickshell IPC channel, not via shell or systemd. The orphan
auditor grep can't see that consumer, so the script trips
audit-orphan-scripts.sh --strict.

Add it to the "Desktop UI entrypoints" category alongside the existing
quickshell-try and hyprshell-trigger entries, with a comment noting the
QS IPC consumer.

---------

+++ kitty/shaders/darkwindow/dither.glsl
@@ -6,7 +6,7 @@
 // CC BY-NC-SA 4.0 (https://creativecommons.org/licenses/by-nc-sa/4.0/)
 
 // Packed bayer pattern using bit manipulation
-const float bayerPattern[4] = float[4](
+const int bayerPattern[4] = int[4](
     0x0514, // Encoding 0,8,2,10
     0xC4E6, // Encoding 12,4,14,6
     0x3B19, // Encoding 3,11,1,9
@@ -15,7 +15,7 @@ const float bayerPattern[4] = float[4](
 
 float getBayerFromPacked(int x, int y) {
     int idx = (x & 3) + ((y & 3) << 2);
-    return float((int(bayerPattern[y & 3]) >> ((x & 3) << 2)) & 0xF) * (1.0 / 16.0);
+    return float((bayerPattern[y & 3] >> ((x & 3) << 2)) & 0xF) * (1.0 / 16.0);
 }
 
 #define LEVELS 2.0 // Available color steps per channel
